fix(table): Fix focusout of Date fields
When selecting a date in the datepicker, the field's <input> loses focus. Moreover, the datepicker exists in the root of the DOM, so is not hierarchically related to the input/control, and is not focusable (no button/tabindex). Thus, on focusout, we can't check for the hierarchical relationships between the element that lost focus and the one that gained focus.

To correctly check when focus is moved to the datepicker, we use the builtin datepicker.inFocus boolean value. Then, we wait for the datepicker to change the input's value to trigger the onControlBlur callback.